+++ src/spectrum.jl
@@ -73,6 +73,7 @@ function spectrum(
     !isendomorphic(H.dimensions) && _non_endomorphic_dims_error("Hamiltonian or Liouvillian for spectrum", H.dimensions)
 
     L = liouvillian(H, c_ops)
+    (L isa QuantumObject) || throw(ArgumentError("spectrum only supports (time-independent) QuantumObject in c_ops"))
     check_mul_dimensions(L, A)
     check_dimensions(A, B)