+++ src/steadystate.jl
@@ -1,4 +1,4 @@
-export steadystate, steadystate_fourier, steadystate_floquet
+export steadystate, steadystate_fourier
 export SteadyStateSolver,
     SteadyStateDirectSolver,
     SteadyStateEigenSolver,
@@ -9,11 +9,18 @@ export SteadyStateSolver,
 abstract type SteadyStateSolver end
 
 @doc raw"""
-    SteadyStateDirectSolver()
+    SteadyStateDirectSolver(
+        factorization = lu
+    )
 
 A solver which solves [`steadystate`](@ref) by finding the inverse of Liouvillian [`SuperOperator`](@ref) using the standard method given in `LinearAlgebra`.
+
+# Arguments
+- `factorization::Function=lu`: The factorization method to use for solving the linear system. The default is `lu`, which performs LU factorization.
 """
-struct SteadyStateDirectSolver <: SteadyStateSolver end
+Base.@kwdef struct SteadyStateDirectSolver{FT <: Function} <: SteadyStateSolver
+    factorization::FT = lu
+end
 
 @doc raw"""
     SteadyStateEigenSolver()
@@ -24,19 +31,29 @@ struct SteadyStateEigenSolver <: SteadyStateSolver end
 
 @doc raw"""
     SteadyStateLinearSolver(
-        alg = KrylovJL_GMRES(; precs = (A, p) -> A isa SparseMatrixCSC ? (ilu(A, τ = 0.01), I) : (I, I))
+        alg = KrylovJL_GMRES(; precs = (A, p) -> A isa SparseMatrixCSC ? (ilu(A, τ = 0.01), I) : (I, I)),
+        ρ0 = nothing,
+        return_details = Val(false),
     )
 
 A solver which solves [`steadystate`](@ref) by finding the inverse of Liouvillian [`SuperOperator`](@ref) using the `alg`orithms given in [`LinearSolve.jl`](https://docs.sciml.ai/LinearSolve/stable/).
 
 # Arguments
 - `alg::SciMLLinearSolveAlgorithm=KrylovJL_GMRES()`: algorithms given in [`LinearSolve.jl`](https://docs.sciml.ai/LinearSolve/stable/)
+- `ρ0::Union{Nothing, QuantumObject}=nothing`: The initial guess of the `steadystate` solution. If not specified, the initial guess will be handled by the solver.
+- `return_details::Val{<:Bool}`: Whether to return the details from the `SciMLBase.LinearSolution`. If `Val(true)`, the [`steadystate`](#ref) function will return a 2-element tuple: the steady state and an extra `NamedTuple` containing the linear solve details. If `Val(false)`, only the steady state will be returned. Default to `Val(false)`.
 
 # Note
 Refer to [`LinearSolve.jl`](https://docs.sciml.ai/LinearSolve/stable/) for more details about the available algorithms. For example, the preconditioners can be defined directly in the solver like: `SteadyStateLinearSolver(alg = KrylovJL_GMRES(; precs = (A, p) -> (I, Diagonal(A))))`.
 """
-Base.@kwdef struct SteadyStateLinearSolver{MT <: Union{SciMLLinearSolveAlgorithm, Nothing}} <: SteadyStateSolver
+Base.@kwdef struct SteadyStateLinearSolver{
+        MT <: Union{SciMLLinearSolveAlgorithm, Nothing},
+        ST <: Union{Nothing, QuantumObject},
+        DT <: Val,
+    } <: SteadyStateSolver
     alg::MT = KrylovJL_GMRES(; precs = (A, p) -> A isa SparseMatrixCSC ? (ilu(A, τ = 0.01), I) : (I, I))
+    ρ0::ST = nothing
+    return_details::DT = Val(false)
 end
 
 @doc raw"""
@@ -45,7 +62,8 @@ end
         ρ0 = nothing,
         tmax = Inf,
         terminate_reltol = 1e-4,
-        terminate_abstol = 1e-6
+        terminate_abstol = 1e-6,
+        return_details = Val(false),
     )
 
 An ordinary differential equation (ODE) solver for solving [`steadystate`](@ref). It solves the stationary state based on [`mesolve`](@ref) with a termination condition.
@@ -65,9 +83,10 @@ or
 # Arguments
 - `alg::AbstractODEAlgorithm=DP5()`: The algorithm to solve the ODE.
 - `ρ0::Union{Nothing,QuantumObject}=nothing`: The initial state of the system. If not specified, a random density matrix state will be generated.
-- `tmax::Real=Inf`: The final time step for the steady state problem.
-- `terminate_reltol` = The relative tolerance for stationary state terminate condition. Default to `1e-4`.
-- `terminate_abstol` = The absolute tolerance for stationary state terminate condition. Default to `1e-6`.
+- `tmax::Real`: The final time step for the steady state problem. Default to `Inf`.
+- `terminate_reltol`: The relative tolerance for stationary state terminate condition. Default to `1e-4`.
+- `terminate_abstol`: The absolute tolerance for stationary state terminate condition. Default to `1e-6`.
+- `return_details::Val{<:Bool}`: Whether to return the details from the ODE solution. If `Val(true)`, the [`steadystate`](#ref) function will return a 2-element tuple: the steady state and an extra `NamedTuple` containing the ODE solve details. If `Val(false)`, only the steady state will be returned. Default to `Val(false)`.
 
 !!! warning "Tolerances for terminate condition"
     The terminate condition tolerances `terminate_reltol` and `terminate_abstol` should be larger than `reltol` and `abstol` of [`mesolve`](@ref), respectively.
@@ -80,12 +99,14 @@ Base.@kwdef struct SteadyStateODESolver{
         TT <: Real,
         RT <: Real,
         AT <: Real,
+        DT <: Val,
     } <: SteadyStateSolver
     alg::MT = DP5()
     ρ0::ST = nothing
     tmax::TT = Inf
     terminate_reltol::RT = 1.0e-4
     terminate_abstol::AT = 1.0e-6
+    return_details::DT = Val(false)
 end
 
 @doc raw"""
@@ -138,34 +159,66 @@ function steadystate(
     return _steadystate(L, solver; kwargs...)
 end
 
-function _steadystate(L::QuantumObject{SuperOperator}, solver::SteadyStateLinearSolver; kwargs...)
-    L_tmp = L.data
-    state_dimensions = L.dimensions.to.op_dims
-    N = get_size(state_dimensions)[1]
-    weight = norm(L_tmp, 1) / length(L_tmp)
-
-    v0 = _dense_similar(L_tmp, N^2)
-    fill!(v0, 0)
-    allowed_setindex!(v0, weight, 1) # Because scalar indexing is not allowed on GPU arrays
-
-    idx_range = collect(1:N)
-    rows = _dense_similar(L_tmp, N)
-    cols = _dense_similar(L_tmp, N)
-    vals = _dense_similar(L_tmp, N)
+# this function is used for SteadyStateDirectSolver and SteadyStateLinearSolver to handle the trace preserving condition
+function _handle_steady_state_trace_preserving_condition(A::AbstractMatrix, N_op::Int, N_super::Int)
+    # [ solving linear problem: Ax = b ]
+    # handle A (Liouvillian) to include the trace preserving condition
+    weight = norm(A, 1) / length(A)
+    idx_range = collect(1:N_op)
+    rows = _dense_similar(A, N_op)
+    cols = _dense_similar(A, N_op)
+    vals = _dense_similar(A, N_op)
     fill!(rows, 1)
-    copyto!(cols, N .* (idx_range .- 1) .+ idx_range)
+    copyto!(cols, N_op .* (idx_range .- 1) .+ idx_range)
     fill!(vals, weight)
-    Tn = _sparse_similar(L_tmp, rows, cols, vals, N^2, N^2)
-    L_tmp = L_tmp + Tn
+    Tn = _sparse_similar(A, rows, cols, vals, N_super, N_super)
+    A_new = A + Tn
 
-    (haskey(kwargs, :Pl) || haskey(kwargs, :Pr)) && error("The use of preconditioners must be defined in the solver.")
+    # handle b to include the trace preserving condition
+    b = _dense_similar(A_new, N_super)
+    fill!(b, 0)
+    allowed_setindex!(b, weight, 1) # Because scalar indexing is not allowed on GPU arrays
 
-    prob = LinearProblem{true}(L_tmp, v0)
-    ρss_vec = solve(prob, solver.alg; kwargs...).u
+    return A_new, b
+end
 
-    ρss = reshape(ρss_vec, N, N)
-    ρss = (ρss + ρss') / 2 # Hermitianize
-    return QuantumObject(ρss, Operator(), state_dimensions)
+function _steadystate(L::QuantumObject{SuperOperator}, solver::SteadyStateLinearSolver; kwargs...)
+    (haskey(kwargs, :Pl) || haskey(kwargs, :Pr)) && throw(ArgumentError("The use of preconditioners (Pl or Pr) must be defined in the solver."))
+
+    # handle dimensions
+    state_dimensions = L.dimensions.to.op_dims
+    N_op = get_size(state_dimensions)[1]
+    N_super = N_op^2
+
+    # handle initial guess of u (steady state) and return dimensions
+    # u0 can be useful for parameter sweeps when the steady state changes smoothly with the parameters
+    u0 = if isnothing(solver.ρ0)
+        nothing
+    else
+        _, u0_data, _, _ = _handle_init_state_and_sol_type_dims(L, solver.ρ0)
+        u0_data
+    end
+
+    # linear problem: Au = b
+    A, b = _handle_steady_state_trace_preserving_condition(L.data, N_op, N_super)
+
+    prob = LinearProblem{true}(A, b, u0 = u0)
+    sol = solve(prob, solver.alg; kwargs...)
+
+    ρss_data = reshape(sol.u, N_op, N_op)
+    ρss = QuantumObject((ρss_data + ρss_data') / 2, Operator(), state_dimensions) # Hermitianize
+
+    if getVal(solver.return_details)
+        # several details stored in SciMLBase.LinearSolution
+        details = (
+            retcode = sol.retcode,
+            stats = sol.stats,
+        )
+        return ρss, details
+    else
+        return ρss
+    end
+    return ρss
 end
 
 function _steadystate(L::QuantumObject{SuperOperator}, solver::SteadyStateEigenSolver; kwargs...)
@@ -182,27 +235,17 @@ function _steadystate(L::QuantumObject{SuperOperator}, solver::SteadyStateEigenS
 end
 
 function _steadystate(L::QuantumObject{SuperOperator}, solver::SteadyStateDirectSolver)
-    L_tmp = L.data
+    # handle dimensions
     state_dimensions = L.dimensions.to.op_dims
-    N = get_size(state_dimensions)[1]
-    weight = norm(L_tmp, 1) / length(L_tmp)
+    N_op = get_size(state_dimensions)[1]
+    N_super = N_op^2
 
-    v0 = _dense_similar(L_tmp, N^2)
-    fill!(v0, 0)
-    allowed_setindex!(v0, weight, 1) # Because scalar indexing is not allowed on GPU arrays
+    # linear problem: Au = b
+    A, b = _handle_steady_state_trace_preserving_condition(L.data, N_op, N_super)
 
-    idx_range = collect(1:N)
-    rows = _dense_similar(L_tmp, N)
-    cols = _dense_similar(L_tmp, N)
-    vals = _dense_similar(L_tmp, N)
-    fill!(rows, 1)
-    copyto!(cols, N .* (idx_range .- 1) .+ idx_range)
-    fill!(vals, weight)
-    Tn = sparse(rows, cols, vals, N^2, N^2)
-    L_tmp = L_tmp + Tn
-
-    ρss_vec = L_tmp \ v0 # This is still not supported on GPU, yet
-    ρss = reshape(ρss_vec, N, N)
+    F = solver.factorization(A)
+    ρss_vec = F \ b
+    ρss = reshape(ρss_vec, N_op, N_op)
     ρss = (ρss + ρss') / 2 # Hermitianize
     return QuantumObject(ρss, Operator(), state_dimensions)
 end
@@ -231,7 +274,16 @@ function _steadystate(L::AbstractQuantumObject{SuperOperator}, solver::SteadySta
 
     sol = mesolve(L, ρ0, tlist; kwargs3...)
     ρss = sol.states[end]
-    return ρss
+
+    if getVal(solver.return_details)
+        details = (
+            retcode = sol.retcode,
+            t_final = sol.times_states[end], # sol.times_states relates to sol.t in OrdinaryDiffEq.jl
+        )
+        return ρss, details
+    else
+        return ρss
+    end
 end
 
 _steadystate(
@@ -326,9 +378,6 @@ In the case of `SSFloquetEffectiveLiouvillian`, instead, the effective Liouvilli
 !!! note "different return"
     The two solvers returns different objects. The `SteadyStateLinearSolver` returns a list of [`QuantumObject`](@ref), containing the density matrices for each Fourier component (``\hat{\rho}_{-n}``, with ``n`` from ``0`` to ``n_\textrm{max}``), while the `SSFloquetEffectiveLiouvillian` returns only ``\hat{\rho}_0``. 
 
-!!! note
-    `steadystate_floquet` is a synonym of `steadystate_fourier`.
-
 ## Arguments
 - `H_0::QuantumObject`: The Hamiltonian or the Liouvillian of the undriven system.
 - `H_p::QuantumObject`: The Hamiltonian or the Liouvillian of the part of the drive that oscillates as ``e^{i \omega t}``.
@@ -361,6 +410,8 @@ function steadystate_fourier(
     check_dimensions(H_0, H_p, H_m)
 
     L_0 = liouvillian(H_0, c_ops)
+    (L_0 isa QuantumObject) || throw(ArgumentError("steadystate_fourier only supports (time-independent) QuantumObject in c_ops"))
+
     L_p = liouvillian(H_p)
     L_m = liouvillian(H_m)
     return _steadystate_fourier(L_0, L_p, L_m, ωd, solver; n_max = n_max, tol = tol, kwargs...)
@@ -445,6 +496,3 @@ function _steadystate_fourier(
 
     return steadystate(L_eff; solver = solver.steadystate_solver, kwargs...)
 end
-
-# TODO: Synonym to align with QuTiP. Track https://github.com/qutip/qutip/issues/2632 when this can be removed.
-const steadystate_floquet = steadystate_fourier