+++ test/core-test/utilities.jl
@@ -25,7 +25,7 @@
         @test n_thermal(0, ω2) == 0.0
         @test n_thermal(ω1, 0) == 0.0
         @test n_thermal(ω1, -ω2) == 0.0
-        @test n_thermal(ω1, ω2) == 1 / (exp(ω1 / ω2) - 1)
+        @test n_thermal(ω1, ω2) ≈ 1 / (exp(ω1 / ω2) - 1)
         @test typeof(n_thermal(Int32(2), Int32(3))) == Float32
         @test typeof(n_thermal(Float32(2), Float32(3))) == Float32
         @test typeof(n_thermal(Int64(2), Int32(3))) == Float64